[Prover Service] Move test-only RSA code into test directory.
