+++ prover-service/src/input_processing/mod.rs
@@ -4,4 +4,3 @@ pub mod field_check_input;
 pub mod field_parser;
 pub mod input_signals;
 pub mod public_inputs_hash;
-pub mod rsa;
